BATT2 addition
Implemented a BATT2 feature (must be enabled under features section in
MavLink_FrSkySPort.ino). This will create a second VFAS sensor using
sensor instance 15. This must be renamed by editing the discovered
sensor to keep the name unique.
This has been tested for compile but not in practise as needs Pixhawk
2.1 with 2 power bricks.